Separated the schedule into it's own component so it can be easily switched out on the page
