+++ frontend/src/Schedule.css
@@ -1,0 +1,206 @@
+#schedule   {
+    background-color: white;
+    float: clear;
+    border: 1px solid #D1D2D4;
+    margin-top: 25px;
+    width: 80%;
+    
+}
+
+#scheduleContainer  {
+    overflow-y: scroll;
+    overflow-x: hidden;
+    max-height: 80vh;
+}
+
+.day    {
+    width: 13%;
+    float: left;
+    text-align: center;
+    border-right: 1px solid #D1D2D4;
+    position: relative;
+}
+
+.dayLabel   {
+    width: 13%;
+    float: left;
+    text-align: center;
+    line-height: 20px;
+    color: white;
+    height: 40px;
+    background-color: #00A9D8;
+}
+
+#hourColumn    {
+    width: 5%;
+    float: left;
+    border-right: 1px solid #D1D2D4;
+    text-align: center;
+}
+
+.hour   {
+    height: 80px;
+    background-color: white;
+    line-height: 40px;
+    vertical-align: middle;
+    border-bottom: 1px solid #D1D2D4;
+    border-top: 1px solid white;
+}
+
+.halfHour   {
+    height: 40px;
+    background-color: white;
+    line-height: 80px;
+    vertical-align: middle;
+    border-bottom: 1px solid #D1D2D4;
+}
+
+::-webkit-scrollbar {
+    width: 15px;
+  }
+  
+  /* Track */
+  ::-webkit-scrollbar-track {
+    background: #f1f1f1; 
+  }
+   
+  /* Handle */
+  ::-webkit-scrollbar-thumb {
+    background: #888; 
+    border-radius: 7px;
+  }
+
+  #days {
+      height: 40px;
+      width: 100%;
+      background-color: #00A9D8;
+  }
+
+  #topRow   {
+      width: 5%;
+      height: 40px;
+      float: left;
+  }
+
+  .appointment  {
+      height: 82px;
+      width: 100%;
+      background-color: #00529b;
+      position: absolute;
+      top: 82px;
+      text-align: left;
+      color: white;
+      overflow: hidden;
+      transition: 0.5s;
+      border-radius: 10px;
+      -webkit-box-shadow: inset 0px 0px 15px 0px rgba(0,0,0,0.2);
+     -moz-box-shadow: inset 0px 0px 15px 0px rgba(0,0,0,0.2);
+     box-shadow: inset 0px 0px 15px 0px rgba(0,0,0,0.2);
+     z-index: 2;
+  }
+
+  .appointment:hover    {
+      height: auto;
+      cursor: pointer;
+  }
+
+  .name {
+      font-weight: bold;
+      font-size: 1.1em;
+      margin-top: 10px;
+      margin-left: 10px;
+  }
+
+  .time  {
+    font-size: 0.8em;
+    margin-top: 5px;
+    margin-left: 10px;
+  }
+
+  .notes    {
+    font-size: 0.8em;
+    margin-top: 15px;
+    padding: 10px;
+    display: none;
+  }
+
+  #timeLine {
+    height: 2px;
+    width: 100%;
+    background-color:red;
+    position: absolute;
+    text-align: left;
+    z-index: 3;
+    position: relative;
+  }
+
+  .hidden   {
+      display: none;
+  }
+
+  @media only screen and (max-width: 900px) {
+    #schedule {
+      width: calc(100% - 50px);
+      margin-left: 50px;
+    }
+  }
+
+  #toggle   {
+      width: 280px;
+      margin: auto;
+      display: none;
+  }
+
+  .daytoggle    {
+      float: left;
+      height: 30px;
+      width: 30px;
+      border-radius: 15px;
+      text-align: center;
+      line-height: 30px;
+      margin: 5px;
+      background-color: #A8A9AD;
+      color: rgb(228, 228, 228);
+      font-weight: 600;
+      transition: 0.5s;
+  }
+
+  .daytoggle:hover  {
+      cursor: pointer;
+  }
+
+  @media only screen and (max-width: 550px) {
+    .day    {
+        width: 90%;
+    }
+    .dayLabel   {
+        width: 90%;
+    }
+    .Monday  {
+        display: none;
+    }
+    .Tuesday  {
+        display: none;
+    }
+    .Wednesday  {
+        display: none;
+    }
+    .Thursday  {
+        display: none;
+    }
+    .Friday  {
+        display: none;
+    }
+    .Saturday  {
+        display: none;
+    }
+    #SundayToggle   {
+        background-color: #EA7600;
+    }
+    #hourColumn {
+        width: 9%;
+    }
+    #toggle {
+        display: block;
+    }
+  }
